tests: merge redundant cli.main test fixtures
