+++ tests/cli/main/test_version_check.py
@@ -3,46 +3,6 @@ from unittest.mock import Mock
 import pytest
 
 import streamlink_cli.main
-from streamlink.logger import capturewarnings
-from streamlink.session import Streamlink
-
-
-# TODO: merge duplicate fixtures from related test modules
-
-
-@pytest.fixture(autouse=True)
-def argv(request: pytest.FixtureRequest, monkeypatch: pytest.MonkeyPatch):
-    argv = getattr(request, "param", [])
-    monkeypatch.setattr("sys.argv", ["streamlink", *argv])
-
-    return argv
-
-
-@pytest.fixture(autouse=True)
-def _setup(monkeypatch: pytest.MonkeyPatch, session: Streamlink):
-    monkeypatch.setattr("streamlink_cli.main.CONFIG_FILES", [])
-    monkeypatch.setattr("streamlink_cli.main.streamlink", session)
-    monkeypatch.setattr("streamlink_cli.main.setup_streamlink", Mock())
-    monkeypatch.setattr("streamlink_cli.main.setup_plugins", Mock())
-    monkeypatch.setattr("streamlink_cli.main.setup_signals", Mock())
-    monkeypatch.setattr("streamlink_cli.argparser.find_default_player", Mock())
-
-    level = streamlink_cli.main.logger.root.level
-
-    try:
-        yield
-    finally:
-        capturewarnings(False)
-        streamlink_cli.main.logger.root.handlers.clear()
-        streamlink_cli.main.logger.root.setLevel(level)
-        streamlink_cli.main.args = None  # type: ignore[assignment]
-        streamlink_cli.main.console = None  # type: ignore[assignment]
-
-
-@pytest.fixture(autouse=True)
-def _euid(request: pytest.FixtureRequest, monkeypatch: pytest.MonkeyPatch):
-    euid = getattr(request, "param", 1000)
-    monkeypatch.setattr("os.geteuid", Mock(return_value=euid), raising=False)
 
 
 @pytest.mark.parametrize(
